frame allocations use new types

+++ include/mem/frame.h
@@ -20,12 +20,12 @@
 
 void frame_init();
 
-uint8_t frame_get(uint32_t frame);
-void frame_set(uint32_t frame, uint8_t status);
+uint8_t frame_get(uintptr_t frame);
+void frame_set(uintptr_t frame, uint8_t status);
 
-uint32_t frame_alloc(uint32_t* frame);
-uint32_t frame_alloc_mult(uint32_t* frames, uint32_t num);
-uint32_t frame_alloc_mult_contig(uint32_t* frames, uint32_t num);
-void frame_free(uint32_t frame);
+uint32_t frame_alloc(uintptr_t* frame);
+uint32_t frame_alloc_mult(uintptr_t* frames, size_t num);
+uint32_t frame_alloc_mult_contig(uintptr_t* frames, size_t num);
+void frame_free(uintptr_t frame);
 
 #endif /*FRAME_H*/
